requirements.txt
ollama
sentence-transformers
faiss-cpu
python-dotenv
numpy
pyyaml
streamlit
